# Peken-Banyumasan-Gate-Masuk-Keluar/openapi-gate.yaml
openapi: 3.1.0
info:
  title: Peken Banyumasan — Gate (Admin & Petugas) API
  version: 2.4.1
  description: |
    API consumed by the Gate admin/petugas panel
    (`Peken-Banyumasan-Gate-Masuk-Keluar/frontend/`).

    **Auth (v2.0 — Supabase):**
    - Login/logout/me/changePassword are **Supabase-direct** — frontend calls
      `supabase.auth.signInWithPassword()` etc. without touching this backend.
    - Gate has **no self-register** — admin and petugas accounts are created by superadmin.
    - `PUT /api/auth/profile` (custom fields only) is **HYBRID**.
    - All other endpoints require a valid **Supabase JWT** in `Authorization: Bearer <token>`.
      Backend verifies using `SUPABASE_JWT_SECRET` and enforces `app_metadata.role IN ('admin', 'petugas')`.
    - Endpoints exclusive to `role: admin` are noted in their summary. Petugas is limited to
      dashboard, visitor scan, and basic event/zone read.

    **Vocabulary (CANONICAL — non-negotiable):**
    - `kolaborator` (NEVER member, kreator)
    - `artisan` (NEVER tenant, umkm)
    - `aktivitas` for admin-moderation story feed (NEVER admin/stories)
    - Path segment `/kolaborator` singular (NEVER `/kolaborators`)
    - Path segment `/artisan` singular (NEVER `/artisans`)

    ---

    **Response Envelope (ALL endpoints except binary file downloads):**
    ```
    Success: { "status": "success", "message": "<optional>", "data": <payload> }
...
    ```
...

    **Authentication:**
...
    does NOT need to implement those endpoints.

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/petugas:
    get:
      operationId: listPetugas
      tags: [petugas-management]
      summary: List all petugas (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [aktif, disabled]
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Petugas list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Petugas'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createPetugas
      tags: [petugas-management]
      summary: Create new petugas account (admin only)
      description: |
        Creates Supabase Auth user with `app_metadata.role = 'petugas'` and
        inserts a `users_profile` row via the existing `trg_sync_user_role` trigger.
        The trigger handles the users_profile insert — backend only needs to call
        Supabase Auth Admin `createUser`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nama, email, password]
              properties:
                nama:
                  type: string
                  example: Budi Santoso
                email:
                  type: string
                  format: email
                  example: budi@pekenbanyumasan.id
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: Initial password — petugas should change via /settings after first login.
                jabatan:
                  type: string
                  nullable: true
                  example: Petugas Pengunjung
      responses:
        '201':
          description: Petugas created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Petugas'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/petugas/{id}:
    get:
      operationId: getPetugas
      tags: [petugas-management]
      summary: Get petugas detail (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Petugas record
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Petugas'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: patchPetugas
      tags: [petugas-management]
      summary: Update petugas profile — nama, jabatan, email (admin only)
      description: |
        Admin can edit: nama, jabatan, email.
        NOT status (use /status endpoint) — status is exclusively managed via
        /api/petugas/{id}/status.
        Email update requires backend to also update auth.users.email.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nama:
                  type: string
                jabatan:
                  type: string
                  nullable: true
                email:
                  type: string
                  format: email
                  description: Updates users_profile.email mirror; backend MUST also update auth.users.
            example:
              nama: Budi Santoso
              jabatan: Koordinator Pintu Masuk
      responses:
        '200':
          description: Updated petugas
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Petugas'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/petugas/{id}/status:
    patch:
      operationId: patchPetugasStatus
      tags: [petugas-management]
      summary: Enable or disable petugas account (admin only)
      description: |
        Disable: sets `banned_until = '2099-12-31T23:59:59Z'` in Supabase Auth
        (native Auth pattern — no DB column needed).
        Enable: sets `banned_until = null`.
        Disabled petugas cannot log in; their users_profile row is preserved.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [aktif, disabled]
            example:
              status: disabled
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                          status:
                            type: string
                            enum: [aktif, disabled]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/petugas/{id}/reset-password:
    post:
      operationId: resetPetugasPassword
      tags: [petugas-management]
      summary: Admin-trigger password reset for petugas (admin only)
      description: |
        Two modes:
        - `email_link`: Supabase Auth sends a reset-link email to the petugas.
          Response returns a confirmation message only (no secret exposed).
        - `temp_password`: Backend generates a secure random password, sets it
          via Admin SDK, and returns it **once** in the response for the admin
          to relay to the petugas. The petugas should change it immediately
          via `/settings`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode:
                  type: string
                  enum: [email_link, temp_password]
            example:
              mode: temp_password
      responses:
        '200':
          description: Password reset initiated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            description: Present for mode=email_link
                          temp_password:
                            type: string
                            description: Present for mode=temp_password — one-time display only
              examples:
                email_link:
                  value:
                    status: success
                    message: Link reset password telah dikirim
                    data:
                      message: Link reset password telah dikirim ke email petugas.
                temp_password:
                  value:
                    status: success
                    message: Password sementara berhasil dibuat
                    data:
                      temp_password: Xk7!mR2qP9
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Artisan Management ────────────────────────────────────────────────────────
...
  /api/zones:
    get:
      operationId: listZonesGlobal
      tags: [zones]
      summary: List global venue zone layout (admin only)
      description: Returns zone layout without per-event occupied state.
      responses:
        '200':
          description: Zone list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ZoneGlobalLayout'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      operationId: saveZonesGlobal
      tags: [zones]
      summary: Replace global venue zone layout (admin only)
      description: Replaces entire layout. Backend validates no occupied stand is removed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [zones]
              properties:
                zones:
                  type: array
                  items:
                    $ref: '#/components/schemas/ZoneGlobalLayout'
      responses:
        '200':
          description: Layout saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
              example:
                status: success
                message: Layout zona berhasil disimpan
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/events/{id}/zones:
    get:
      operationId: listEventZones
      tags: [zones]
      summary: List zone layout with per-event occupied state (admin + artisan)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Zone layout with occupied state
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ZoneEventOccupancy'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase JWT (RS256). Issued by `supabase.auth.signInWithPassword()` on the frontend.
        Backend verifies using `SUPABASE_JWT_SECRET`.
        Role must be 'admin' OR 'petugas' in `app_metadata.role`.
        Endpoints marked "admin only" additionally reject role 'petugas' with 403.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource UUID or prefixed short ID
      example: evt-0001
  responses:
    Unauthorized:
      description: Missing, invalid, or expired Supabase JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Sesi Anda telah berakhir
            data: null
    Forbidden:
      description: Authenticated but role insufficient (e.g. petugas accessing admin-only endpoint)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Akses ditolak
            data: null
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Sumber daya tidak ditemukan
            data: null
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Validasi gagal
            data: null
            errors:
              status: [nilai tidak valid]
  schemas:
    Envelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        message:
          type: string
          nullable: true
        data:
          description: Actual payload — shape varies per endpoint
    ErrorEnvelope:
      type: object
      required: [status, message, data]
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
        data:
          type: object
          nullable: true
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    Petugas:
      type: object
      required: [id, nama, email, status, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: auth.users.id
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        nama:
          type: string
          description: users_profile.nama
          example: Budi Santoso
        jabatan:
          type: string
          nullable: true
          description: users_profile.jabatan
          example: Petugas Pengunjung
        email:
          type: string
          format: email
          description: auth.users.email
          example: budi@pekenbanyumasan.id
        status:
          type: string
          enum: [aktif, disabled]
          description: Derived from auth.users.banned_until — disabled if banned_until is set and in the future.
          example: aktif
        created_at:
          type: string
          format: date-time
          description: UTC ISO 8601 — auth.users.created_at
          example: '2026-01-15T08:00:00.000Z'
        last_sign_in_at:
          type: string
          format: date-time
          nullable: true
          description: UTC ISO 8601 — auth.users.last_sign_in_at
          example: '2026-05-01T10:30:00.000Z'
...
    ZoneGlobalLayout:
      type: object
      description: |
        Static venue layout (used by GET/PUT /api/zones).
        Reflects DB zones table: stands array stores {id, label, kategori} only.
      required: [zona, label, warna, stands]
      properties:
        zona:
          type: string
          example: A
        label:
          type: string
          example: Zona A – Kriya & Fashion
        warna:
          type: string
          example: '#8B5E3C'
        urutan:
          type: integer
          description: Display order
          example: 1
        stands:
          type: array
          items:
            type: object
            required: [id, label, kategori]
            properties:
              id:
                type: string
                example: A-1
              label:
                type: string
                example: Stand A-1
              kategori:
                type: string
                example: Kriya
    ZoneEventOccupancy:
      type: object
      description: |
        Per-event zone occupancy (used by GET /api/events/{id}/zones).
        Backend joins zones with event_artisans WHERE status_request='approved'
        OR pending (for orange overlay in FE ZoneSelector).
      required: [zona, label, warna, stands]
      properties:
        zona:
          type: string
          example: A
        label:
          type: string
          example: Zona A – Kriya & Fashion
        warna:
          type: string
          example: '#8B5E3C'
        stands:
          type: array
          items:
            type: object
            required: [id, occupied]
            properties:
              id:
                type: string
                example: A-1
              label:
                type: string
              kategori:
                type: string
              occupied:
                type: boolean
                description: True if stand has approved or pending artisan
              artisan_id:
                type: string
                description: Present when occupied
              nama_usaha:
                type: string
                description: Present when occupied
              status_request:
                type: string
                enum: [approved, pending, pending_change]
                description: For UI pending overlay (orange in ZoneSelector)
...